+++ addon/helpers/-private/get-outlet-state.js
@@ -2,9 +2,10 @@ import { setInternalHelperManager } from '@glimmer/manager';
 import { createComputeRef } from '@glimmer/reference';
 
 export default setInternalHelperManager(function (args, owner, scope) {
-  const outletState = scope?.outletState?.compute();
-
   return createComputeRef(() => {
+    // args; owner; scope; debugger
+    const outletState = scope?.outletState?.compute?.();
+
     return outletState || null;
   });
 }, {});