+++ tests/dummy/app/components/can-import.svelte
@@ -1,7 +1,0 @@
-<script>
-  import Hello from './hello.svelte';
-</script>
-
-<Hello>
-  <h3 data-test-can-import>Imports a Svelte component</h3>
-</Hello>