+++ tests/dummy/app/components/module-context.svelte
@@ -1,3 +1,0 @@
-<script context="module">
-  export let message = 'Ember is Omakase';
-</script>