+++ tests/dummy/app/components/tagged.svelte
@@ -1,3 +1,0 @@
-<svelte:options tag="tagged-component" />
-
-<h2><slot></slot></h2>