+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow use of Svelte components as route templates
+- Support invoking Ember components within Svelte components
 ### Modified
 - Replace document fragment hack with direct Glimmer integration for default slot
+- Replace AST transformation in favor of a runtime approach
+- Fix bug in the build process that could cause file to be written to the actual project directory
 
 ## [0.1.1] - 2022-07-01
 ### Modified